adding required_migrations_checks to BaseCommand to support unapplied migrationgs on other managements command, runserver, createsuperuser and changepassword activate this check
